+++ lib/dnsCache.js
