+++ lib/session.js
@@ -113,7 +113,7 @@ Session.prototype.close = function () {
     this.opened = false;
     this.session.close();
     this.readWatcher.stop();
-    // TODO - remove listeners so program will exit I guess?
+    this.removeAllListeners();
 };
 
 Session.prototype.stats = function () {
